Add feature that allows auto-typing to libvirt using virDomainSendKey.
This change extends the auto-type implementation to allow multiple
auto-type plugins to live alongside the platform-native plugins, and
the auto-type to libvirt is implemeneted with such a plugin.

The difference with the platform-native plugins is that these
'external' plugins execute the auto-type on a selected target (in
the case of the libvirt plugin a libvirt domain UUID is used to
identify the target). The plugins provide a list of available
targets, which are presented to the user as target selection.