Add dead keys support in libvirt plugin and refactor AutoTypeActions.
When guests targeted by the libvirt plugin use a keyboard layout
with dead keys, certain characters need a trailing space keycode to
properly output the character. Because Windows seems to be the only
OS that uses dead keys in the default keyboard layout (US-intl),
two new options added to the settings page. One to enable dead keys
if a Windows guest is detected (on by default), and the other to
enable dead keys for other guest operating systems (off by default).

The AutoTypeActions refactor replaces the targetIdentifier string
with a AutoTypeTarget object.

Plugins can implement AutoTypeTarget subclasses to share context
between listing available targets and performing the auto-type
action. For the libvirt plugin it is used to keep a libvirt pointer,
so that it doesn't have to fetch the domain and operating system
information for every sent char/key.