Implement history purge feature in Database Maintenance
Fixes #7550

The database maintenance widget is complemented with a section that allows the user to clean up old history entries. the age of history entries to be removed is configurable. respective entries are searched for and removed with the click of a button. message boxes inform about whether or not and if so how many entries were removed from the database. Reuse the `maintenanceHistoryDays` attribute from KeePass2 to store the desired days of history to maintain.