Remove bogus fallback
It if XDG_RUNTIME_DIR is unset, then QStandardsPaths::RuntimeLocation
returns fallback a path in /tmp. This implies that the fallback code
remove here is wrong, since the fallback to /tmp is already handled by
Qt and doesn't need to be done again on the KeePassXC side.

It's clear that nobody is relying on the existing fallback for the
extension, since the code-path that creates triggers this fallback was
unreachable.