Fix Ctrl+Tab shortcut to cycle databases in unlock dialog
When in PR #8168 the abbility to cycle through the databases in the unlock
dialog was extended to include the `Ctrl+Tab / Ctrl+Shift+Tab` shortcuts,
it mixed up the direction. At the moment `Ctrl+Tab` selects the previous
database, when customarily it should select the next entry. This was my
mistake.

This commit changes the direction of cycling through databases in the
unlock dialog with `Ctrl+Tab / Ctrl+Shift+Tab` fixing my original commit.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>