Simplify Q_OS_* macro logic
Nested `#ifdef` are easy to get wrong, especially when inner ones do
not indent (for example `# ifdef ONE_LEVEL_DEEP_MEANS_ONE_SPACE`).

Flatten out the three branches for MacOS, Windows and UNIX and wrap the
later MacOS specific code again.

`defined()` must be used here as `#elifdef` is not a thing and `#elif`
cannot be used as it requires the macro to be defined, which is not how
`Q_OS_*` macros work -- this little but important detailed got pointed
out by KeePassXC's CI tests after submitting this PR as draft.