Disable AES-256/GCM for now
Currently, the granularity for the botan gcm implementation is too large.
To fix a problem with another algorithm in the library, they are multiplying
the blocksize, so by default the granularity is 64. This causes issues since
the encrypted data in the key is only guaranteed to have a length that is a
multiple of the block size (16).