Add optional support for Botan3
Update CMake config to add a WITH_XC_BOTAN3 option. This will result in
CMake setting the C++ standard to C++20 instead of C++17, ignoring the
'deprecated-enum-enum-conversion' and 'deprecated' warnings, and using
the Botan 3 version/include dir/libraries instead of the Botan 2 ones.

The value of this option will be made available to the c++ code via the
config-keepassxc.h header.

This also makes some minimal compatability fixes to the code (behind a
ifndef-else-endif compiler directive)

- For Botan 3, just check that the botan major version is 3
- In SessionCipher.cpp, if we are compiling against Botan 3, use the
  pubkey api instead of the pk_ops api.