Implement autosaveDelay in DatabaseWidget
Implement a autosave delay functionality to the database widget.
Added timer to track the delay and trigger a save on timeout using the new onAutosaveDelayTimeout function.
Changed the onDatabaseModified to test for the new autosaveDelay property and delay the save if needed or extend the delay on new modifications.
After menual save stop the timer to avoid saving when not needed edited in DatabaseWidget::save().

Part of issue keepassxreboot/keepassxc#8553