Allow configuring shortcuts
Closes #2689

This pull request aims to introduce shortcut configuration to keepassxc.

The design of the respective code is loosely based on KDE's KActionCollection. The ActionCollection manages all actions that can be shortcut configured. These actions are then exposed in the config and a user can assign a different shortcut.

Actions inside the MainWindow have been added to the ActionCollection.

The GUI for configuring actions is pretty simple, it just consists of a QTreeView and QKeySequenceEdit to change the shortcut.