Merger: prevent duplicate entry when merging histories
If the alien database (source) entry is newer, a copy of the entry is
made. But before moving the merged entry to the target group, it must be
removed. Otherwise there will be briefly two entries with the same UUID
in the same group/database.

Even though this is only the case during the transaction, it can still
be observed because the operations emit signals. A notable problem is
the fdosecrets feature that relies on the uniqueness of the UUID or will
otherwise run into problems because the UUID is used as part of the DBus
path.