Fix Copy Password button when text is selected
When the user chooses to copy the password for an entry to the
clipboard, previously there was logic to check if text was selected, and
if so, that text was instead copied to the clipboard. That made sense if
(a) the user invoked the Copy Password action via its keyboard shortcut,
and (b) that keyboard shortcut was configured (as per default) to be
Ctrl-C, i.e. the same as the system action for copy-to-clipboard.

However, it made no sense if the user invoked that action in some other
way, for example by clicking the corresponding toolbar button.

It also made no sense in the case that the Copy Password action had some
other keyboard shortcut assigned. Also, if some other action had Ctrl-C
assigned, the logic would not kick in then.

Fix all of the above by modifying the keyboard shortcut logic to
intervene precisely in the case where a shortcut is pressed that matches
the system copy-to-clipboard shortcut; only in that case do we now check
if text is selected and if so copy that to the clipboard instead of the
action we would otherwise take.

Add a test case to TestGui.cpp.

Fixes #10734.