EditEntryWidget: SSH Agent: Load destination constraints into temporary settings
When using the "Add to agent" button in the SSH Agent category of the
entry settings a default KeeAgentSettings structure will be used and
filled-in with the current state of the UI elements. As destination
contraints are not present in the UI for now those have not been loaded
and thus the key was loaded into the SSH agent with no constraints at
all. This change fixes it by loading the attachment first before
overwriting with the data from the UI elements.

Signed-off-by: Konrad Vité <kvite@paktolos.net>