Avoid deprecated operator+(QMap::iterator, int)
This operator+ is deprecated since Qt6.2:
https://doc.qt.io/archives/qt-6.2/qmap-iterator.html.

QMap is implemented as a red/black tree, presumably without storing
subtree sizes, so this is a linear lookup.  We need to call this for
each displayed alias (multiple times), so linear complexity for lookup
by rank is a bummer.  But presumably nobody's going to have that many
aliases, that this actually matters.

Suggested-by: Copilot