Add FdoSecrets/CollectionAliasDatabaseUUIDs config
Add a setting to store the Freedesktop Secret Service aliases.
Setting e.g. the "default" alias indicates, to which database most
applications should store new secrets.

This is a QVariantMap (aka QMap<QString, QVariant>) as opposed to a
QMap<QString, QUuid>, as only the former can be stored to QSettings.
Converting a QVariantMap to QMap<QString, QUuid> would require copying
the entire map, calling QVariant::toUuid on every value.

This commit does not yet add the user interface for configuring this
setting, nor does it  apply the configured aliases to the actual D-Bus
service.