Adding support for theme-aware tray icon in KDE/Plasma
If a KDE/Plasma session is detected, offer to use a monochrome but theme-aware icon
for the system tray. Here, 'theme-aware' means that the SVG icon is augmented so
that the KDE SVG renderer can replace the main color to fit the current theme, e.g.
use a light color if the theme is dark.