Do not translate backup filename TIME placeholder format
The format string should be written using the tokens derived from
English, i.e. "yyyy" for years. The example leads one to believe that
tokens derived from the user's current language could be used. Fix
Finnish and Ukrainian because those are the two that I could spot that
make this error.